ttools: generalise TableRowHeader for use with huge tables
Fix TableRowHeader so it doesn't propagate TableModelEvents unnecessarily,
and add a hook to alllow you to configure the column width.
Also fix it so it can display longs rather than ints (not used yet, but
no reason not to).

+++ topcat/src/main/uk/ac/starlink/topcat/ColumnInfoWindow.java
@@ -377,7 +377,7 @@ public class ColumnInfoWindow extends AuxWindow {
 
         /* Set up a row header. */
         TableRowHeader rowHead = new TableRowHeader( jtab ) {
-            public int rowNumber( int irow ) {
+            public long rowNumber( int irow ) {
                 return irow;
             }
         };
